pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "google-genai>=1.38.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "reportlab>=4.4.3",
]
